Might've fixed minima/maxima.

+++ Cercetare/Structs.h
@@ -2,18 +2,18 @@
 
 struct Centroid
 {
-	unsigned x;
-	unsigned y;
-	unsigned z;
-	unsigned sum_x;
-	unsigned sum_y;
-	unsigned sum_z;
+	float x;
+	float y;
+	float z;
+	float sum_x;
+	float sum_y;
+	float sum_z;
 	int count;
 };
 
 struct Neuron
 {
-	unsigned x;
-	unsigned y;
-	unsigned z;
+	float x;
+	float y;
+	float z;
 };
